requirements.txt
Flask==2.3.2
Flask_SQLAlchemy==3.1.1
PyMySQL==1.1.0